auth/requirements.txt
setuptools==46.4.0
django-oauth-toolkit
djangorestframework
django
django-phonenumber-field[phonenumbers]
drf-pretty-exception-handler
pytest-drf
pytest-django
pytest
behave
behave-django
django-grpc-framework
grpcio==1.39.0
grpcio-tools